Extract globals and reused properties
