+++ src/components/react_multi_select.scss
@@ -1,0 +1,116 @@
+@import "../resources/globals";
+
+.wrapper {
+  display: flex;
+  position: relative;
+  min-width: $space-920;
+
+  .list_container {
+    width: 50%;
+  }
+
+  .destination_header {
+    display: flex;
+    justify-content: space-between;
+    border: $border-1 solid $grey-350;
+    border-width: $border-1 $border-1 0 0;
+    height: $space-750;
+    padding: 0 $space-400;
+    align-items: center;
+    font-size: 12px;
+
+    .clear_all {
+      color: $grey-600;
+      cursor: pointer;
+    }
+  }
+
+  .loader_container {
+    position: absolute;
+    width: calc(100% - 2px);
+    height: calc(100% - 2px);
+    background-color: $color-white;
+    left: 1px;
+    z-index: 99;
+  }
+
+  .checkbox_control {
+    height: $space-700;
+  }
+
+  .search_icon {
+    right: $space-400;
+    font-size: $font-400;
+    line-height: $lineheight-14;
+    color: $grey-600;
+  }
+
+  .source_items_wrapper {
+    height: $space-930;
+  }
+
+  .source_item {
+    height: $space-700;
+    line-height: $lineheight-40;
+    padding: 0 $space-400;
+    width: 100%;
+
+    &:hover.selected_source_item, &:hover:not(.selected_source_item) {
+      background-color: $grey-200;
+    }
+
+    &.selected_source_item {
+      background-color: $grey-100;
+    }
+
+    & > span {
+      width: 100%;
+    }
+  }
+
+  .remove_selected_icon {
+    margin-right: $space-200;
+    margin-top: $space-70;
+  }
+
+  .destination_items_wrapper {
+    height: $space-940;
+    border-left: 0;
+  }
+
+  .destination_item {
+    height: $space-700;
+    line-height: $lineheight-40;
+    padding: 0 $space-100;
+    transition: none;
+
+    &:hover:not(.selected_destination_item) {
+      background-color: $grey-200;
+      cursor: pointer;
+
+      .destination_item_content {
+        color: $grey-600;
+
+        .remove_button {
+          color: $color-white;
+        }
+      }
+    }
+
+    .destination_item_content {
+      display: flex;
+      justify-content: space-between;
+
+      .destination_item_text {
+        overflow: hidden;
+        text-overflow: ellipsis;
+        margin-right: $space-400;
+      }
+
+      .remove_button {
+        font-size: $font-400;
+      }
+    }
+  }
+
+}
