+++ src/components/react_multi_select_utils.js
