--- src/react_multi_select.scss
@@ -1,115 +1,0 @@
-.wrapper {
-  display: flex;
-  position: relative;
-  min-width: 420px;
-
-  .list_container {
-    width: 50%;
-  }
-
-  .selected_header {
-    display: flex;
-    justify-content: space-between;
-    border: 1px solid #DDDDDD;
-    border-width: 1px 1px 0 0;
-    height: 51px;
-    padding: 0 20px;
-    align-items: center;
-    font-size: 12px;
-
-    .clear_all {
-      color: #666666;
-      cursor: pointer;
-    }
-  }
-
-  .loader_container {
-    position: absolute;
-    width: calc(100% - 2px);
-    height: calc(100% - 2px);
-    background-color: #ffffff;
-    left: 1px;
-    z-index: 99;
-  }
-}
-
-.checkbox_control {
-  height: 40px;
-}
-
-.search_icon {
-  right: 20px;
-  font-size: 14px;
-  line-height: 14px;
-  color: #666666;
-}
-
-.source_items_wrapper {
-  height: 323px;
-}
-
-.source_item {
-  height: 40px;
-  line-height: 40px;
-  padding: 0 20px;
-  width: 100%;
-
-  &:hover.selected_source_item, &:hover:not(.selected_source_item) {
-    background-color: #f3f3f3;
-  }
-
-  &.selected_source_item {
-    background-color: #f7f7f7;
-  }
-
-  & > span {
-    width: 100%;
-  }
-}
-
-.remove_selected_icon {
-  margin-right: 12px;
-  margin-top: 7px;
-}
-
-
-
-.dst_items_wrapper {
-  height: 363px;
-  border-left: 0;
-}
-
-.dst_item {
-  height: 40px;
-  line-height: 40px;
-  padding: 0 10px;
-  transition: none;
-
-  &:hover:not(.selected_dst_item) {
-    background-color: #f3f3f3;
-    cursor: pointer;
-
-    .dst_item_content {
-      color: #666666;
-
-      .remove_button {
-        color: #ffffff;
-      }
-    }
-  }
-
-  .dst_item_content {
-    display: flex;
-    justify-content: space-between;
-
-    .dst_item_text {
-      overflow: hidden;
-      text-overflow: ellipsis;
-      margin-right: 20px;
-    }
-
-    .remove_button {
-      font-size: 14px;
-    }
-  }
-}
