+++ src/resources/globals.scss
@@ -1,0 +1,35 @@
+// Fonts
+$font-400: 14px;
+$font-300: 12px;
+
+// Spacing
+$space-940: 360px;
+$space-930: 323px;
+$space-920: 300px;
+$space-860: 180px;
+$space-750: 50px;
+$space-700: 40px;
+$space-400: 20px;
+$space-200: 12px;
+$space-100: 10px;
+$space-70: 7px;
+
+// Borders
+$border-1: 1px;
+
+// Line heights
+$lineheight-14: 14px;
+$lineheight-40: 40px;
+
+//Colors
+$color-white: #ffffff;
+$color-black: #000000;
+$grey-350: #dddddd;
+$grey-450: #999999;
+$grey-420: #b4b4b4;
+$grey-200: #f3f3f3;
+$grey-600: #666666;
+$grey-100: #f7f7f7;
+
+$blue-100: #D9ECF1;
+
